PP-5655 White list naxsi rule 1002 for stripe notifications
Do not block Stripe notifications when the `id` field contains `0x`. According
to the documentation "...for JSON, naxsi does not keep track of depth" and given
the example it provides this rule should apply to any field named 'id'. It is
possible there can be multiple fields named 'id' and none should be checked by
rule 1002.

Link to naxsi json documentation:
https://github.com/nbs-system/naxsi/wiki/json

Associated Zendesk ticket is: 3807008